Minor style adjustments over much of the site.
Heavily reworked cards on home page so they shrink "gracefully", though it's
still a little weird at intermediate viewports. Images are now color-coded
with the buttons they go to, with the "conservation" card now leading to the
red list, which is not actually implemented (it goes to the "about" page).